Make `fetch::refs::update()` private again, move tests accordingly. (#450)
It's so hard to call I doubt it should ever be called manually outside
of the confines of its test-suite.
