+++ git-repository/src/remote/fetch.rs
@@ -2,8 +2,9 @@ use crate::bstr::BString;
 
 /// If `Yes`, don't really make changes but do as much as possible to get an idea of what would be done.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum DryRun {
+pub(crate) enum DryRun {
     /// Enable dry-run mode and don't actually change the underlying repository in any way.
+    #[cfg(test)]
     Yes,
     /// Run the operation like normal, making changes to the underlying repository.
     No,
